+++ med-booking-frontend/src/index.tsx
@@ -4,7 +4,7 @@ import { Provider } from "react-redux";
 import { BrowserRouter } from "react-router-dom";
 import { PersistGate } from "redux-persist/integration/react";
 import App from "./App";
-import "./index.css";
+import "./css/index.css";
 import { persistor, store } from "./store/store";
 
 const root = ReactDOM.createRoot(
